Add sitemap generation
We need to generate a sitemap that google uses to process CPC codes.
Since the sitemap may be costly to generate, we don't want to do it
dynamically but instead will create a new one on a regular basis.

This adds a package.json hook to the sitemap code, which can later be
triggered by cron (or inn our case Heroku scheduler) on whatever cadence
we want.

The code is based off of work done in the v1 repository but refactors to
use async functions instead of promises, as this is much easier to read
and follow the flow.

Issue #32