Manage Lambda error responses better
The response we receive when invoking our `queryParser()` Lambda
function is a little unusual. If there’s an error in calling the
function (such as if we call an non-existent function name), the
`invoke()` callback function supplies an `err` value.

But if the function itself throws an error, that comes back with a clean
`StatusCode: 200` but with the `data.Payload` being a “true” object (not
a JSON string) with an `errorMessage` attribute.

So now, we:

1. First, single-parse `data.Payload`
2. Then, check for `data.Payload.errorMessage`
3. If it exists, reject the promise with a stringified `data.Payload`
   (for inspection)
4. If it doens’t exist, continue on, but parse `data.Payload` a second
   time, since it presumably contains the valid JSON string that
   `queryParser()` returned.

Affects #16